Cambios en organización de estructura del proyecto
